requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.34
Django==2.2.13
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-eml-email-backend==0.1
django-environ==0.4.5
django-js-asset==1.2.2
django-rq==2.3.2
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
natsort==7.0.1
packaging==20.4
Pillow==7.1.2
pipdeptree==1.0.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
rq-scheduler==0.10.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
